Moved predefined templates
